tests/requirements.txt
pytest==7.4.0
requests==2.31.0
pytest-html==3.2.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
minio==7.2.0